⚙️ Make LaserCannon defs reflect real range
This is NOT a balance change! These units already had such range.
Spring requires LaserCannon range to be an integer multiple of
projectile velocity and lowers it if need be. The new values are
what the previous values were shrunk to.

The best way to see this effect may be Newton, because it has a
convenient toggle to switch between two weapons with different
range but keeping everything else (such as unit position, emit
points etc) the same. The attractive blue ray was supposed to be
longer ranged than the repulsive red ray, but it was not larger
enough to reach more frames, therefore they get rounded to the
same value. You can easily check that the blue ray has the exact
same range as the red one.

Added a brutal check to make sure that anybody touching balance
is aware of this effect, similar to frame-bound reload times.

Chicken Leaper had its projectile velocity increased instead to
demonstrate how to make the range match nominal, if an increase
in range less than a single frame's worth of velocity was needed.

Heatrays still calculate damage reduction based on the old range,
since that keeps previous behaviour and I don't want this change
to touch balance. Anything other than Scorcher probably wouldn't
be affected too much if they started using the new range though.