🐛 Fix transports picking up moving enemies
The first allyteam to build a transport would have all their units
pickable up by their enemies despite moving (because that transport's
script environment would linger and be used for allyteam checks).

This "ugly hax" is probably not going anywhere (since it's good that
the engine lets games decide) so this sounds like a nice opportunity
to say that it closes #2775.