Lagmonitor:  * AFK time 30 -> 45 -> 40  * Warn time 5 -> 8  * Units with BP are set to wait when shared, wait is removed upon return.
Giving a factory an order that cancels its rally queue also cancels the wait. This is consistent with cancelling wait on other unit types. Without this system the only way to cancel wait would be to actually issue a wait command - requiring knowledge of Wait - which is something I'm trying to avoid.