🐛 Fix #4457, Raven dive vs Aegis
Shield dive works by having the fake weapon get intercepted by shields and
prompting a dive. There are two components to the failure:

 * there is a distance check which prevents diving too early.
   In the past, this check was ignored for shields (no distance
   was passed). af798ae added distance to one of the checks.

 * the first time the shield check happens is just right for it
   to be stopped via the above. The weapon fires every 1s so the
   next check is then too late. 06f6c23 reduced the frequency.

I decided to approach this via the first component and removed the
distance check. This seemed less scary and also makes the code agree
with the comments inside.