Added a command to build a plate to the factory tab in spot BZ aka row 3, column 1. When command is issued, cursor shows a plate icon, and if cursor moves in range of a factory, the active command gets switched to building a plate of that factory.
It's implemented as a generic do-nothing command in LuaRules, and I edited cmd_factory_plate_placer.lua in LuaUI to add the behavior there in a way that does not disrupt the current plating behavior.  This allows it to also use the LuaUI glsl stuff in the current widget, which is familiar and looks nicer.