⚖️ Zenith fire once now fires just 1/3 of meteors
Tries to alleviate the main issue of the unit (which is that meteors past
the first ~30 are mostly useless since everything except Det dies anyway)
without ruining the "firing spends way too much ammo" aspect of the unit.

This is a new mechanical ability for the unit and not just a bit of UI,
unlike Fire Once for other units. This is somewhat important not just
because of efficiency but also because overflow is bad so if you could
send a small, fixed amount of meteors into a corner it would dilute this
penalty.