Added new units to CAI and adjusted existing ones
For jumps: Placeholder as riot, replacing Jack.
For spider: Venom as raider, Crab as arty.
For shield: Felon as riot, Rack as arty.
For hover: Claymore as riot.
For rover: Crasher as aa.
For amph: Bulkhead as arty, Limpet as riot (testing CAI with mobile bombs).
For planes: Likho and Thunderbird as bombers.
For gunships: Locust as raider (replacing Blastwing), Trident as fighter, Revenant and Krow as bombers (not sure about Krow, we'll see).
For turrets: Gauss, at a 3:7 ratio with stinger.