HeadsUp function added
potential fix for #4636
HeadsUp function moves the 'head' of lance's weapon upwards so that
it doesn't accidentally friendly fire anything as it wobbles randomly
downward. Model was not ment to bend at the base of the crane, so it
looks a little strange. May be better to change wobble behavior instead.
Lance also liable to hit friendlies with spherical hitboxes as they pass
in front due to firing downwards at most things on level terrain.