✨ Mexspot labels now rotate smoothly
Rotating the camera now rotates the labels smoothly. The expensive thing
is remaking the GL list, so when the camera doesn't rotate the check can
be done as often as we want since it's just a numerical function call.
I couldn't detect a performance drop on rotation either.

Math was also simplified a bit thanks to the new engine API.