🐛 Fix #4798, faux precision in Context Menu
Looks like calls of the function already had some sort of argument
which the signature of the function didn't know about and which got
"activated" with the recent speed change.

Clear it as I can't tell what it was (probably precision, but in
that case some places where it was applied didn't make sense) and
it was applied fairly haphazardly anyway.