⚙️ Reclaim highlight doesn't count mixed energy
Used to count 70 energy as 2 metal, on the grounds that Solar produces
at that ratio. But that doesn't really work since Solar produces the
other way around and also the +2 is per second, not a one-time chunk.

Then there was the inconsistency where energy-only features (trees etc)
weren't counted, so that conversion ratio didn't apply anyway, while
wreckage never has energy. So this only applied to map features with
mixed contents which are fairly rare.