⚙️ Ctrl automorph: disable, works with Shift
Too unpolished to be enabled by default. See #4806

Works with Shift again (this effectively reverts the earlier
commit de53b37c5e4003aff273ef999b0253c0bebf8f03).

This means that queuing a rectangle will again morph all the
structures, but this will now only affect those who enabled
the feature and it is more important that single queue works.

Ideally there would be smart detection of single vs rectangle,
at which point it could be enabled by default, but that is not
in the scope of this patch.