⚙️ Remove redundant `requireammo` customparam
Now the `reammoseconds` param is necessary and sufficient to have
a rearming bomber. All bombers have the field and in that case it's
better not to have a separate "enabled" param and a default value
for the reload, since then all code needs to be aware of the default
and apply it if the main param is present and the other is missing.