Magpie seems to have won the name competition.
Bolas has 4.5% more DPS and can 1-shot Flea.
 * Reload 0.366 -> 0.433
 * Damage 34 -> 42

Phoenix is heavier and has its secret weapon rework documented.
 * Cost 360 -> 460
 * Health 1060 -> 1450 (90 more than if scaling by cost)
 * Reduced spray angle slightly
 * Drops per second halved to 7.5 Hz
 * Area of Effect 216 -> 320
 * Burn time 10s -> 12s
 * Damage per bomb 25 -> 50
It deals about 3x more damage in testing against a carpet of light units (10k -> 30k vs Ravager), with about 10% more damage to units in the hottest part of the run (about 550). The main change to the weapon is a run that reaches twice as far due to the halved bomb drop rate.