Followed "Fight your opponent, not the UI" through to removing manual groundcrash technique from Odin.  * The micro involves waiting until Odin is flying over the target, then firing, for 800 or so extra damage.  * Implemented via enforced minimum horizontal velocity.
The removal of the physics interaction and leverageable knowledge is sad, but probably for the best. Alternates might be:
 * Reduce shot range significantly, so it always groundcrashes. This would look a bit silly, and also be a pain because Odin would have to line up with its target perfectly.
 * Some sort of UI for selecting groundcrash - even worse.
 * Delve beyond "Fight your opponent, not the UI", but that doesn't seem warranted for a one-off mechanic.

With this in place it seems to be impossible to deal more than 8050 damage by shooting at Grizzly. However, there is even now still a technique that 1-shots it. But this technique is disrupted by prettymuch any unit bumping up against the Grizzly, although it will be possible to accidentally 1-shot your own Grizzly by dodging very slightly into the right spot for the technique.