Disco Rave Party:  * Cost 42k -> 45k  * Range 7000 -> 9600 (Zenith is 8800)  * Projectile gravity 0.18 -> 0.1 It now comfortably outranges Zenith but pays a bit more for the privilege. If a Zenith is scouted then building a DRP just beyond its range should be viable. This seems good since DRP is much worse vs armies due to its turn rate, especially flanking armies. The significantly lower gravity makes the difference between high and low trajectory more extreme. High trajectory has much higher spread due to much higher flight time, so DRP is now unique among superweapons in that it much prefers to be built with a free line of fire towards the enemy.
Starlight deforms the map at 1/5th the rate, with about 5x the power, to work around the unreasonably high performance cost of such deformation that seems to be a recent thing in the engine. It looks a bit janky, especially on deforming to water, but it is better than people suddenly lagging out.