Here is how I would tweak it
- Reduced particle size to help with the jarring 180 turns.
- Changed final red 1,0,0 -> 1,0.2,0.2 to remove the soft-red effect. Now it looks like a "heat thing" until the end.
- Reduced initial size to reflect low DPS
- Increased final spread for the finale
- Also tweaked launch tail and no-damage white.

It looks a little bit weak at the start, good for the middle 15s or so, then ridiculous in the last few seconds.I think this is a good place to be as the first few seconds teach the mechanic and the shift at the end can be used to gauge when it is over. Something that deals 80 DPS shouldn't look massive.