⚙️ Add an unfinished support powers gadget
Backend for casting Age of Mythology / CnC Generals style global powers.
Initially meant for Planet Wars, but will surely be of interest for modders
and was made with Spring/Recoil basecontent in mind to some extent.
It's an unfinished WIP so contains no GUI, potentially questionable API
design decisions, and a lot of FIXMEs. There's some talk of reviving PW
so perhaps somebody might want to take it up from there.

To check it out:
* enable via modoption: `all_support_powers=1`
* cast a tacnuke at coordinates: `/luarules support tacnuke X Z`