⚙️ Allow regular on/off on impulse units
Currently on/off always gets turned into push/pull on impulse units.

But you could have a unit with a single mode (say, just push) and an
unrelated on/off ability (say, jammer) in which case it shouldn't be
turned into push/pull. An example would be a modded commander.

Add a weapondef customparam, `notoggleimpulse`, which makes a unit
use regular on/off if all its impulse weapons are tagged with it.

Feature stolen from FW. Thanks Shaman!