⚙️ Gadgeted sound now considered battle SFX
Battle sounds spawned via the gadget interface (like Djinn teleport
loop, or rez sound) are now played in the "battle" channel like other
in-world sounds.

As far as I can tell the only consequence is that it is controlled by
a different volume slider, but that's still good.

Stolen from FW. Thanks Shaman!