Siren shoots a bit faster.  * Reload 1.7 -> 1.6 -> 1.066  * Damage 280 -> 175 This is the same DPS as pre-buff Siren, but ends up being a lone TTK buff against many of its targets.
The most notable are a nerf vs Cutter (1-shot to 2-shot) and a decent buff vs Seawolf (5.1s to 4.3s to kill). In practise, its large AoE will confound these calculations. Eg previous Siren got Seawolf down to 14% health in two shots, while the new one gets it down to 21% health in 3 shots. So previous Siren had a bit of a better shot at using AoE to steal a Seawolf for 2 direct shots. I still expect new Siren to steal Seawolves in 3 direct shots in actual play though.

A not-insignificant part of this change is feel. Siren at 1.7s reload time spends a lot of time doing nothing. There is no reload bar because the missile takes that slot in the UI. The gun takes about 0.9s to finish its recoil (and it already does it quite slowly) and the effects of the sonic projectile are over in about 0.7s. So Siren would spend almost half its reload time doing nothing but aiming, which is frustrating because you unit can be dying and "doing nothing about it". I noticed this after feeling how much more active the lower reload time Siren feels.