🐛 Fix orbital dropping ships
Ships wanted to land on the seafloor like all other units, but there
was water in the way preventing them from doing so, so they never got
outside the "falling" state. There is a failsafe to catch this but
for xponen reasons it checked whether the unit is going upwards (as
opposed to just not falling). The less lazy fix would be to make ships
target the surface instead of the seafloor.