✨ Likho remodel. (#5164)
* ✨ Likho remodel.

* Clamp new Likho teamcolour (ie alpha) to 1 or 0 for consistency with the rest of the game.
There are probably arguments to be made for blending teamcolour into the main body of models, but no other unit in ZK does this, and I recall it being fixed as an error back in the heyday of modelling. I think we are better off maintaining these bits of consistency rather than breaking them.

* Tweak peak reflections down by 20%.

* ✨ Likho remodel.

* ✨ Next revision of Likho remodel.

* Save old Likho for comparison.

* Update PR with this zip that is on the thread.

* Fixed Likho script piece names.

* Clamp new Likho teamcolour (ie alpha) to 1 or 0 for consistency with the rest of the game.
There are probably arguments to be made for blending teamcolour into the main body of models, but no other unit in ZK does this, and I recall it being fixed as an error back in the heyday of modelling. I think we are better off maintaining these bits of consistency rather than breaking them.

* Tweak peak reflections down by 20%.

* Remove redundant files that probably appeared during conflict hell.
Tweaked lups.
Fixed model offset.

* Fix new Likho piece hierarchy and lups ribbon.

* Tweak ceg and offset.

* Tweak.

---------

Co-authored-by: bruno-cb <bruno.calmon@hotmail.com>
Co-authored-by: GoogleFrog <googl3frog@gmail.com>