Try a Revenant change  * Speed 135 -> 130. This makes it more vulnerable to raiders, give a little more warning and time under turrets, slows down the rush slightly, and makes it cost a bit more to switch between fronts. A big part of Revenant seems to be its ability to suddenly create a 2v1 situation.  * Damage 220 -> 210. Revenant was buffed over a year ago, so why did it show up in the past few months? It previously did 3960 damage in two shots, so that it 3-shot factories, but this is enough damage to 2-shot an Engineer comm, even with self-regen. It might be that Revenant only became meta since the rise in engineer popularity. This change prevents two Revenants from killing Engineer in one shot by themselves.  * Reload 9 -> 10, to make it worse when caught out by raiders. Revenant rushes on comms etc seem to rarely shoot on cooldown, so it isn't as large change as it looks.
Health seems like a do-nothing nerf in most early cases, since Revenants rarely go super low, but will nerf mid-lategame Revenant. Early game seems like the problem, so I'd avoid doing things that mostly hit other phases.
Cost is an option, but seems too general.
The most worrying aspect is speed, because while it is a 3.7% nerf, the power of Revenant lies in how much faster it is than raiders that would try to get in its way, so zero speed is not the best comparison. This change brings it 11% closer to the speed of the slowest raiders.