rework of gauss_fire and gauss_hit for different units
Rework of gauss sounds and separation of large gauss from small gauss.

hoverraid and gaussturret now have a custom shortened gauss_fire_small and gauss_hit_small with a smaller footprint that better matches the unit, but still sounds like gauss because it's generated from the same base sample. This freed up gauss_fire and gauss_hit to be huge impactful sounds as needed for detriment and cloaksnipe.