More explosion balance
Fixed the new ex_large18 from stereo to mono. returned the presence for Ex_large4 which is now the shogun's impact sound. Will need to figure out quake/raveparty later as it doesn't make sense for either of those to be sharing Ex-Large4 with artillery. Minimal change to ex_large5 from first pass. Swapped ex_large8 with the basic band pass but it appears to be unused.