Update heavy_laser6.wav
"Phantom impact sound is hard to describe. It feels flat but also has too many details." "Lance sounds really weird. Flatness and other effects. It doesn't sound like it is synchronised to the visuals the way the previous sound is. Possibly the previous sound ramps up a bit until the end then drops off." "Lucifer has the Lance problem." I returned to a light touch on this. Dialed back the shortening too, but it doesn't make sense for the sound to continue for the original 4 seconds when the firing animation is 1 second. I probably need to take the cloaksnipe impact off of the lance firing sound but I don't yet know what that new sound should be.