Update wolverine_hit.wav
"Firewalker impact point blank into a structure is just broken." I swapped the original impact sound in to test it and this appears to be broken in the base game currently. It's one of those 'layering' effects when too many of the same sounds are played concurrently that sounds bad. I don't see a solution other than changing the amount of impacts the firewalker is allowed to make. I could try to find the frequency range which is being amplified thats making it sound weird and reduce it but that would almost certainly make the normal functioning sound weird and bad. But, since this hasn't been an issue in base game, I doubt it actually needs adjustment. I returned firewalker impact to it's normal length and only the basic band pass for now as it makes more sense to couple with the fire on the ground lingering.