heatray_fire6 rework
Needed to do a full heatray rework to reduce the length. It fires at 10x second but the original legnth is 0.7. so with twin heatrays like on desolator or dante, there ends up being 14 concurrents playbacks of the audio. This is what causes heatray to "consume" the entire battlefield's audio. We're no down to 0.2 length which still means 4 concurrent audio for dante and desolator and 2 for scorcher. But since it's so easy to spam scorcher, it can still run away with the 'layering', especially if a single scorcher has too much impact. I've settled on a balance that I think works, at least much better than base game, where 1 scorcher is still noticeable and 20 scorchers still demands everyone's attention.