Cloak_Snipe overhaul
"- Phantom firing sound is a lot softer than its impact sound. This is a bit weird as they tend to make up the same action, so ends up sounding like a volume shift. The sounds could be closer in volume, with the firing going up more than the impact goes down."

Phantom has been bugging me for a long time. It's too unique to be sharing sounds with another unit, which also frees up detriment and lance to be more specialized. I did an overhaul on both the firing and hit. Overlaid some unique texture and frequency that sells the fantasy of the invisible sniper. I pushed it fairly hard with the same philosophy of the glaive - Loud, Clear, Unique.