Update wolverine_hit.wav
"- Point blank Firewalker is still broken. Spawn a Firewalker like this and hear for yourself. When fired at range, the shots are each individually too crisp, which breaks up the impact into a series of hits rather than the general laying of a carpet of fire from stable."

Default Firewalker is fundamentally broken in a way that is hard to address. The pleasing mud that covers the impacts you're describing are the result of sub 40hz frequencies that make up a huge amount of the firewalker impact. 99% of Sound systems can't produce those frequencies, and instead produce distortion. This distortion is the mud that's working as . The problem is that every sound system is going to handle that distortion and the sounds on top of it differently, and before the sound rework there was so much distortion all the time that it flew under the radar. Using intentional distortion as an effect reliant on the limitations of each persons sound system while hurting the overall soundscape is probably not the route we want to go. There's no good way to quality control a sound like that and everyone will hear it differently. I've found a middle ground in taking the original and only doing a light cut of frequencies below 40hz and a slightly softened attack.