Zenith Impact
"- The Zenith meteor impact sound is too crisp for when a cluster of them hit the ground. The individual impacts by themselves may also be too flat."

Split zenith off from cyclops. Gave it more presence. I'm liking the impact of it but I can soften the attack if needed. This is another one that was using unreliable distortion to smooth out impacts like the firewalker.