Update heavy_laser6.wav
"- Is Lance firing still a bit too short? I'm not sure I can tell at this point. My feeling watching it is that the main part of the sound (not the aftermath) cuts out a fraction of a second before the beam does.
- Lucifier sounds wrong, compared to Lance, and I don't know why. Fixing Lance probably solves it.
The first 0.6 seconds make up the punch of the initial weapon firing. This is the part that the new sound accentuates.
From around 0.55s to 0.9s, the sound of the initial fire is bridged by a region of extreme clipping that sounds a bit like VwaVwaVa when you zoom into it.
At around 1s the "debris flying sound" swells and sustains at that intensity until around 1.3s. This gives the end of the beam its own audio cue."

Recreated the Lance sound to have lots of presence and no falloff until after the beam has fully ended.