Update generic_cannon.wav
"Ripper impact sound seems very low-fidelity compared to the rest. Jarring."

The ripper impact is 8000hz sample which is the sample rate of a landline telephone. It has always been trash. I'm trying to salvage it by layering a better sample underneath so it keeps some of it's distinctness. If this doesn't work then we need to throw away the ripper impact file and source a new sound entirely.