Update blade_swing2.wav
"Scythe still sounds like it has a slash with a more thumpy sound underneath it. Like dropping a large sheet of cardboard on the floor. Things that are a single action shouldn't sound like two actions as that creates event-noise. The new slash alone may work, but that said, the piercing nature of the previous slash is a feature, not a bug. Scythe needs to globally announce itself."\

I was confused before what you meant by 2 sounds. It didn't sound wrong to me before, but I went back and consolidated the sound so it's more of a single event.