Update emp_missile_hit.wav
"The Shockley impact sound could be more noticeable somehow, which probably involves being less flat (like Imp). This sound should be noticeable as it tends to precede something important happening for most players in the game."

Created more density with another layer so that the sound could have a better attack.