Update flak_fire.wav
"Ettin firing has a noticeable poweh sound that sounds like it clips over the rest of the sound, even though it isn't. This creates a clear repeat. The impact is very echoy and also a bit flat at the same time. I think it is an improvement though. The lower fire rate of Thresher makes its flak sound a bit better."

Created a new sound for Ettin and Flak