Update rapid_laser.wav
"Gremlin, Raptor etc still sound very underwhelming. Compare flying planes over a group of Gremlins to flying over a group of missile AA like Vandal or Angler. The plane dying doesn't square with the sound that is made is kill it. Part of it is that it sounds like a "recording studio" sound that isn't integrated into the world somehow. Perhaps a small amount of echo would fix it. Raptor see Gremlin. Hercules see Gremlin.

Created new teal laser. It feels more in line with the other lasers now like flea or bandit.