Update cannon_hit5.wav
"Tremor impacts probably need to be louder or slightly more noticeable in some way (higher pitch?). I know this is dangerous, but I can't relaly tell what I'm hearing when I see a rain of Tremor shots land somewhere. The firing sound also fails to be awesome, but in this case there is a tradeoff between singleplayer awe and multiplayer playability that is hard to satisfy, so the firing sound is probably fine."

Boosted up the Tremor both volume and distinction. I think it's slightly too loud now but it's easy to reign back down.