Ripper impact update
"The new Ripper impact is far too weak. The original impact is what it should sound like, but higher quality/better balanced in some way."

Trying to keep the best of both worlds by blending some of the old sound with a higher quality base so it keeps some of the texture and recognition but doesn't sound so thin now that it's being exposed to a more balanced sound environment.