Update vlaunch_hit2.wav
"The Impaler is ok, but I'm not sure how it is better than the old one. Won't the maraca-type sound over the top of the deeper impact sound more annoying when spammed? There also seems to be extra echo at the end that echos the sound over the top, rather than the impact. This sound can happen a lot so shouldn't have too much going on. It's a bit like my Scythe comment, except that Impaler shots are much spammier and much less important than Scythe. So I'm not sure why it sort of sounds like two sounds put together."

I was already on the fence on this so I’ve reverted to the version without the reverb added. The thinking was that the Impaler is a unit that does 800 damage in a single shot, and sometimes that shot is coming from fog of war. I was looking for a way to let the sound exist a little longer to be caught by the player since a player might only get a handful of audio cues before something devastating happens. It’s a balance since the impaler is spammable late game, but I also don’t want a new player to lose all of their defenses to some impalers within a couple minutes and not know it’s happening.