Envoy Impact
"Envoy impact sound might be a bit too truncated, was it changed at some point?"

Yes, it used to share a sound with a lot of units. I've isolated it's original length to cannon_hit7 so that it's no longer sharing an impact with Duck. Easy to switch back if we want a less noisy envoy.