Starlight Laser
"The Starlight firing sound is in the TARGETER weapon."

Boosted starlight special fire and gave it some differentiation from the lance sound by splitting and adding texture. I'm maxed out on volume. This one is strange in that it always seems quieter than it should, maybe because it's sourced from the structure and not the point it hits. I tested a version in stereo for the double impact that overcomes this and it works, so that's available if we want to lean harder into the "oppressive superweapon" feeling.