Update heavy_laser7.wav
"Moderator is a lot better. I'd want to hear a version without the wet click about halfway through though, as removing it might improve it (or it might make the sound, idk)."

Created this version of moderator. Moderator loses a little distinction but fits the style of the soundscape better.