Teal Laser and Maxconcurrent test for teal laser + nimbus
"Gremlin is quite a bit better but probably a touch too flat. The sound is a lot less clear than other sounds. It works fairly well when a group of Gremlin are firing in semi-random ways. Raptor and Toad have trouble with their regular frequent fire."

There's a lot here. Created a new teal Laser and split it into 2 variants just like scorcher so the slower fire rate (Gremlin 0.3, Hercules 0.3,) can have a separate sister-sound to the rapid firing units (toad 0.10, Raptor 0.10, detriment 0.10)

The big test here is using maxconcurrent in sounds.lua to prevent the situations where the fastest firing units destroy the soundscape by layering volume. I need to play with this more but I'm excited that it's more robust than I expected. The key is to identify what number of simultaneously firing units break the soundscape and go 1 under that in maxconcurrent to prevent only the worst outcomes. Applied to raptor, toad, and nimbus for now as a test.