Dagger / gauss
"Dagger/Gauss have something weird going on. The start of the impact sound is good, but it turns into whistly flat sound at the end, that is almost like an attenuated echo. It gets worse when played multiple times one frame apart, which happens a lot given how gauss works. Also, if possible, it would probably be good to make the rising pitches in the tail of the firing sound quieter."

Reworked the Gauss and dagger. I'm reaching the limits of what I can do with the original source.