Update generic_cannon.wav
"I mainly don't understand what was wrong with the previous Ripper and Impaler.
    • Ripper sounds very gravelly. I am not sure what it is supposed to be. The old boomph-pop fit together and was clear. I think ZK should generally sound a bit stylised/cartoony, and this Ripper impact is far from it."

The base game ripper has an extremely low sample rate of 8000 khz which is very low quality. It became more noticeable once the sound environment was cleaner, you commented about the original "Ripper impact sound seems very low-fidelity compared to the rest. Jarring." My last two attempts were to try to pair the old sound to a new sound to retain some of it's distinct character, but that's not working. We either have to settle for the imperfections of the base game ripper or abandon it to find an entirely new sound. Attempting to create a direct sound-alike isn't going to work since I have to make due with the samples that I have access to.

Restored Ripper to it's original source for now.