Gauss and Dagger volume balance
"Dagger is pretty good. I'd go for more volume on the fire than the hit so it's easier to hear both events. Gauss fire and hit volume could be adjusted too. Gauss has longer range though, so the two sounds don't need the help that Dagger needs to be distinct."

Adjusted these volumes to be more balanced.