Update flak_fire.wav
"Ettin and Thresher are still a bit weird. The firing and impact don't sound like they belong in the same environment. The firing sound is good but it might need to change to match everything else. Maybe add a bit of reverb for the firing sound? Not a lot, maybe something like the level of Ravager."

Gave the ettin thresher firing sound some reverb to match the environment better.