Cerberus and Shogun impacts
"• Cerberus and Shogun are still a bit echoy and lacking in weight. Eg they are higher pitched than Envoy which makes me think lower weight, plus they have more echo. This is a lot better though and much closer to where it should be. We probably can't have these weapons at their full weight given how spammy they are."

Made adjustments to pitch and echo. Not sure if it's better or just different.